API

HEPAnalysisUtils.errorcap_coords — Method

The two functions here are to make yerror customizable.

See also https://github.com/JuliaPlots/Plots.jl/issues/2156

HEPAnalysisUtils.miunitstop — Method

See also MIGrad in Chapter 4: Minuit Commands, MINUIT Reference Manual

HEPAnalysisUtils.optimizeat — Method

optimizeat(f, res0::Optim.OptimizationResults; pois...)

Optimize a function of a NamedTuple-like varialbe, f, where parameters of interest (pois) are fixed (specified as keyword arguments) and nuisances parameter are floated.

HEPAnalysisUtils.poissonerrs — Method

c.f. https://root.cern.ch/doc/v608/TH18cxxsource.html#l08189

HEPAnalysisUtils.profile — Method

profile(f; pois...)

Curry a function of a NamedTuple-like variable.

Examples

julia> f(x) = x.a + x.b + x.c
julia> profile(f; a=1)(b=2, c=3)
6

HEPAnalysisUtils.unzip2tuples — Method

Wait until this settles: https://github.com/JuliaLang/julia/issues/13942

Optim.optimize — Method

So that it uses the Miunit stopping criteria automatically

A bad thing is that it collides with the the default function signature since there is no default_options mechanism for optimize function that takes Fminbox (as per Optim 1.7.4)

See also this issue.